Performance/memory usage improvements
lomb() was taking ~10 seconds to run and, without creating autorelease pools, was running out of memory on device; even with judicious use of autorelease pools, it was peaking at ~350MB due to performing operations with several large matrices of double precision (~64MB each). By reconsidering and simplifying the calculations to avoid use of large matrices, I've gotten this down to ~2 seconds and no significant increase in peak memory use over the 130MB baseline due to having the json objects from which the data is extracted in memory.